Network commands are now eval'd on the server
The code box is now sent to the server and ran in an eval there. This
makes it possible to actually change the server variables and have the
changes be streamed to all clients.

There's also a var called 'player' as a helper for accessing the
player's entity.

player.translateBy(15,-15,0);

for example will move the player serverside and stream that information
back to the client automatically. The following will simulate the
'hyperspace' effect:

var i = 0;
var moveInterval = setInterval(function(){
if(i < 20) {
player.translateBy(15,-15,0);
i++;
}
else {
clearInterval(moveInterval);
}
}, 100);