New Bullet Object and Failed Bullet/Orb Collision
I am trying to get the Orbs to break into smaller Orbs when a Bullet
collides with them--like in Asteroids.

I added a new Bullet object, by copying the Orb object.   I then added
the box2d body component to both the Orb and Bullet objects.
I made the Orb dynamic and the bullet kinematic in the box2d properties.

I also added a collision listener component in the client.

I changed the sample code in the ace editor to fire one bullet.

The collision listener works, but since the Orb is kinematic all the
Orbs do strange things and get reset to x=0 when the Bullet is fired.

I don't think Box2d is the way to go.