Folder structure
The folder structure seems to be copied straight from the examples but
this should not be.  From the ige manual:
folder structure you should aim for will look something like this:

+ Root Development Folder
+-- ige (Where you unzipped or cloned the engine to)
+-- Your game folder (where you do all your game development)

I have no idea what the dev environment in the server end looks like,
but it should match the manual, no? Now the files in this repo do at
least.