Moved to box2d physics
The asteroids now use box2d and collide with things. The ship can also
shoot by pressing space, I haven't looked at the tablet controls yet,
just made the old shoot method to call the new one.

The asteroids are also currently contained in a 500/500 cube to keep
them from escaping to the infinite space.

When shooting the smallest asteroid there is a 9% chance that a new big
one will be spawned, to keep asteroids in the field. If chances are
against you and you manage to shoot down all asteroids, just make a new
one by submitting

new Orb();

in the ace editor.