Added fixedorb, and fixedorbz
When you collide with a fixedorb, it is destroyed and a fixedorbz is
created of a different color.

This needs to be changed so that the color of the fixedorbz that is
created, matches the color of the ship.

The simple game will be to see who can touch the most fixed orbs before
the other player.  Easter egg hunt type mechanic.