Ige loader, codebox update, code cleanup, polish
Modified the loader to load files side-by-side, not one-by-one. Makes
loading times considerable faster when developing, but will have to be
reverted when packaging the game for launch.

Codebox now has console.log transmission, try
console.log(ige.server.players[clientId]);
or a shorthand
log(me);

Implemented a lesson loading system, there's a readme file in
lessons/readme with instructions on how the lesson files work. Ask me if
it's unclear.

Player shape is now dynamic, the texture reads entity.shape array and
draws itself upon that. This variable is also streamed to all clients,
so changes are visible on all clients. The array functions as follows:
[
[x,y],
[x,y],
[x,y],
...]

so, [0,0] would add a point in the middle of the texture, [1,1] in the
lower right corner, and [-1,-1] in the upper left. [entityGeometry * x,
entityGeometry * y] to be specific.

Code cleanup and polish, and a bit more cleanup. Moved functions around
to make the files more clutter free and organised. Planetoids now use
just one class which changes colour when touched, player respawn is in
Player.explode() and Player.respawn()