Database initialisation
The game now requires a mysql database to be up when starting the
server. ServerConfig.js has the config variables for it.
On the default settings, please create a mysql user
'superspace'@'localhost' with the password that's set in ServerConfig.js
and a database 'superspace' that our user has permissions to use.

Permissions for the user required
SELECT, INSERT, UPDATE, DELETE, CREATE, ALTER, INDEX