Tag Level Edits
Created a “Tag” Level for Parsons Playtech event
player shape changes to hexagon when there is contact with another
player
code lessons were updated to give advantages in the tag game such as
increasing thrust, hyperspace to corner, and creating asteroid field
around your self
there was also a function added to the server that is currently
commented out that creates an octopus type boss.  it uses orbs that are
joined together with revolute joints.  I had to change something in one
of the engine files to allow for revolute joints.