Added Blockly functionality
Finally blockly working in an iframe for a basic scaling of the ship.
Then had help from another developer, Jay Bloodworth, who did a great
job building out all the blocks and code generation to create code used
in power_codes for: timing, hyperspace, orb creation, changing the
shape and color of the ship

using the Google blockly library downloaded from:
https://github.com/google/blockly
- I think all the blockly files are up to date

The only files from the large blockly library that we added or are
editing are(I am fairly sure):
starcoder.js, and playershape.js in the blocks folder
playershape.js in the generators/javascript folder
code.js and index.html in the apps/code folder

You can press “h” to hide or show the iframe.

Added this code to the code.js file in apps/code

var code = Blockly.JavaScript.workspaceToCode();

	parentWindow.postMessage(code,"*");

Added code to index.php to receive the message and send the code to the
server.