Added particles, boss, planet, and tree prototype
Lesson 15 creates a planet
Once you shoot enough asteroids to get your score above 100, when you
touch the planet, it creates a new “tree” object
(I still need to attach this tree to the planet with a revolute joint)
The “boss”  / “hydra” is a series of planetoids attached to one center
“head”.  Bullets bounce off of every part except for the head of the
hydra.
I also used the particles from the “50 - lunar lander” example.
I changed the str in Jay’s code for adding a new lesson to be a string.